split classes and add unwrap (test to be finished)
