+++ eigen_utils/include/eigen_utils/eigen_utils.hpp
@@ -1,15 +1,11 @@
-#ifndef __eigen_util_h__
-#define __eigen_util_h__
+#pragma once
 
 #include "eigen_utils_common.hpp"
-//#include "eigen_gl.hpp"
-//#include "eigen_lcmgl.hpp"
 #include "eigen_rand.hpp"
 #include "eigen_rigidbody.hpp"
 #include "eigen_numerical.hpp"
 #include "eigen_select_block.hpp"
-//#include "eigen_lcm.hpp"
 #include "eigen_file_io.hpp"
 
 
-#endif
+
